+++ bin/mamba.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:mamba/executor.dart';
 import 'package:mamba/mamba_cli.dart';
 
-const _mambaVersion = '0.8.1';
+const _mambaVersion = '0.8.2';
 
 Future<void> main(List<String> arguments) => Executor(
   'mamba',